client.py
import asyncio
from typing import Optional
from contextlib import AsyncExitStack

from mcp import ClientSession, StdioServerParameters
from mcp.client.stdio import stdio_client

from anthropic import AsyncAnthropic
from anthropic.types import ToolParam, Message

from dotenv import load_dotenv

load_dotenv()  # load environment variables from .env

class MCPClient:
    def __init__(self):
        # Initialize session and client objects
        self.session: Optional[ClientSession] = None
        self.exit_stack = AsyncExitStack()
        self.anthropic = AsyncAnthropic()
    # methods will go here

    async def connect_to_server(self, server_script_path: str):
        """Connect to an MCP server

        Args:
            server_script_path: Path to the server script (.py or .js)
        """
        is_python = server_script_path.endswith('.py')
        is_js = server_script_path.endswith('.js')
        if not (is_python or is_js):
            raise ValueError("Server script must be a .py or .js file")

        command = "python" if is_python else "node"
        server_params = StdioServerParameters(
            command=command,
            args=[server_script_path],
            env=None
        )

        stdio_transport = await self.exit_stack.enter_async_context(stdio_client(server_params))
        self.stdio, self.write = stdio_transport
        self.session = await self.exit_stack.enter_async_context(ClientSession(self.stdio, self.write))

        await self.session.initialize()

        # List available tools
        response = await self.session.list_tools()
        tools = response.tools
        print("\nConnected to server with tools:", [tool.name for tool in tools])
    async def process_query(self, query: str) -> str:
        """Process a query using Claude and available tools"""
        if self.session is None:
            raise RuntimeError("Session is not initialized. Call connect_to_server() first.")

        # Convert the human message to the SDK MessageParam object
        messages = [{"role": "user", "content": query}]

        # Get available tools from the server and convert to ToolParam for the Anthropic SDK
        response = await self.session.list_tools()
        available_tools = response.tools  # list of tool objects from the MCP server
        tool_params = [ToolParam(**tool.model_dump()) for tool in available_tools]

        # Call Claude (await the async call) using MessageParam objects
        response = await self.anthropic.messages.create(
            model="claude-3-5-sonnet-20241022",
            max_tokens=1000,
            messages=messages,
            tools=tool_params
        )

        # Collect text outputs; safely handle other content types (e.g., tool_use) without raising type errors
        final_text = []
        for content in getattr(response, "content", []):
            text = getattr(content, "text", None)
            if text:
                final_text.append(text)
            else:
                ctype = getattr(content, "type", None)
                if ctype == "tool_use":
                    name = getattr(content, "name", "<tool>")
                    final_text.append(f"[Tool call requested: {name}]")
                else:
                    # Generic fallback for unknown content shapes
                    final_text.append(str(content))

        return "\n".join(final_text) if final_text else ""
        return ""

    async def chat_loop(self):
        """Run an interactive chat loop"""
        print("\nMCP Client Started!")
        print("Type your queries or 'quit' to exit.")

        while True:
            try:
                query = input("\nQuery: ").strip()

                if query.lower() == 'quit':
                    break

                response = await self.process_query(query)
                print("\n" + response)

            except Exception as e:
                print(f"\nError: {str(e)}")

    async def cleanup(self):
        """Clean up resources"""
        await self.exit_stack.aclose()


async def main():
    #if len(sys.argv) < 2:
     #   print("Usage: python client.py <path_to_server_script>")
    #sys.exit(1)

  #  print(sys.argv[1])

    client = MCPClient()
    try:
        await client.connect_to_server('..\\my-first-mcp-server\\main.py')
        await client.chat_loop()
    finally:
        await client.cleanup()

if __name__ == "__main__":
    import sys
    asyncio.run(main())