pyproject.toml
[project]
name = "my-first-mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.71.0",
    "mcp[cli]>=1.18.0",
    "python-dotenv>=1.1.1",
]

